refactor code to build 2 binaries that shares codes
