+++ Cargo.toml
@@ -2,6 +2,7 @@
 name = "elton"
 version = "0.1.0"
 edition = "2021"
+default-run = "web"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -31,3 +32,13 @@ opt-level = 3
 default = ["web"]
 web = []
 cli = []
+
+[lib]
+name = "modules"
+path = "src/lib.rs"
+
+[[bin]]
+name = "cli"
+
+[[web]]
+name = "web"
