+++ src/benchmark.rs
@@ -1,3 +1,5 @@
+use crate::types::{CreateTest, HttpMethods, Test, TestResult};
+use crate::utils::median;
 use anyhow::Result;
 use chrono::Utc;
 use hyper::{Body, Client, Request};
@@ -9,29 +11,7 @@ use tokio::{
     time::{sleep, Instant},
 };
 
-use crate::{
-    routes::{self, HttpMethods},
-    utils::median,
-};
-
-#[derive(Debug, Clone)]
-pub struct Test {
-    pub second: i64,
-    pub error_code: Option<String>,
-    pub response_code: u16,
-    pub response_time: u64,
-}
-
-#[derive(Debug, Clone)]
-pub struct TestResult {
-    pub second: i64,
-    pub error_codes: Vec<String>,
-    pub response_codes: Vec<u16>,
-    pub requests: i64,
-    pub avg_response_time: f64,
-}
-
-pub async fn run_benchmark(test: routes::CreateTest) -> Result<Vec<(i64, Vec<TestResult>)>> {
+pub async fn run_benchmark(test: CreateTest) -> Result<Vec<(i64, Vec<TestResult>)>> {
     log::info!(
         "Starting benchmark using {} tasks for {} seconds",
         test.tasks,
@@ -191,7 +171,7 @@ pub async fn run_benchmark(test: routes::CreateTest) -> Result<Vec<(i64, Vec<Tes
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::routes::{CreateTest, HttpMethods};
+    use crate::types::{CreateTest, HttpMethods};
     use httptest::{matchers::*, responders::*, Expectation, ServerPool};
 
     // Create a server pool that will create at most 2 servers.
