+++ src/bin/cli/main.rs
@@ -1,19 +1,15 @@
 use clap::Parser;
 
-/// Simple program to greet a person
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
 struct Args {
-    /// Name of the person to greet
     #[arg(short, long)]
     name: String,
-
-    /// Number of times to greet
     #[arg(short, long, default_value_t = 1)]
     count: u8,
 }
 
-pub fn run() {
+pub fn main() {
     let args = Args::parse();
 
     for _ in 0..args.count {
